cookie bug in dashboard
i thought i wasnt able to send the cookie back to the server. The reality was that i wasnt creating a cookie from the login and was able to find that when i erased the saved cookie and tried to login again. So in my post req in axios i made a small change and managed to get the data.